templates/backends/python-fastapi/requirements.txt
fastapi[standard]>=0.115.0
uvicorn[standard]>=0.34.0
sqlalchemy[asyncio]>=2.0.36
asyncpg>=0.30.0
alembic>=1.14.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.2.0
pyotp>=2.9.0
qrcode>=8.0
aiosmtplib>=3.0.0
jinja2>=3.1.0
httpx>=0.28.0
python-multipart>=0.0.18
slowapi>=0.1.9
python-dotenv>=1.0.1
